+++ OperationCenter/OperationCenter.h
@@ -13,6 +13,8 @@
 #include "Commons.h"
 #include "DataProcessor.h"
 #include "Parser.h"
+#include "Logic.h"
+#include "Feedback.h"
 
 using namespace std;
 class OperationCenter{
@@ -28,51 +30,6 @@ public:
 	static string showWelcomeMessage(bool status);
 };
 
-class Logic{
 
-private:
-	static const string ADD_COMMAND;
-	static const string DELETE_COMMAND;
-	static const string DISPLAY_COMMAND;
-	static const string CLEAR_COMMAND;
-	static const string SORT_COMMAND;
-	static const string SEARCH_COMMAND;
-	static const string EDIT_COMMAND;
-	static const string EXIT_COMMAND;
-	static const string UNDO_COMMAND;
-	static const string EMPTY_RESPONSE;
-	static const string IVALID_COMMAND_MESSAGE;
-
-
-public:
-	static void executeInput(string input);
-	static void executeCommand(string& returnDisplay, string& returnResponse, string command, Data task, int taskNo, TimeMacro currentTime);
-	static void saveData();
-	static void loadData(bool& status);
-	static string displayIfEmpty(string returnDisplay, TimeMacro current, TimeMacro start, TimeMacro end);
-	static string displayToday(string returnDisplay, TimeMacro current);
-	static TimeMacro setCurrentTime();
-	static string showWelcomeMessage(bool status);
-};
-
-class Feedback{
-private:
-	 static string display;
-	 static string response;
-public:
-
-	static string getDisplay(){
-		return display;
-	}
-	static string getResponse(){
-		return response;
-	}
-	static void updateDisplay(string newDisplay){
-		display = newDisplay;
-	}
-	static void updateResponse(string newResponse){
-		response = newResponse;
-	}
-};
 
 #endif
