+++ OperationCenter/OperationCenter.vcxproj
@@ -69,6 +69,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="Feedback.h" />
+    <ClInclude Include="Logic.h" />
     <ClInclude Include="OperationCenter.h" />
   </ItemGroup>
   <ItemGroup>
