+++ OperationCenter/OperationCenter.vcxproj.filters
@@ -18,6 +18,12 @@
     <ClInclude Include="OperationCenter.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Logic.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Feedback.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="OperationCenter.cpp">
